api/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
psycopg2==2.9.10
python-dotenv==1.0.1
